requirements.txt
faiss-cpu 
sentence-transformers
tf-keras